Crate hopr_chain_connector

Modules

blokli_client

Re-exports of the blokli_client crate.

errors

Structs

Address

Represents an Ethereum address

BlockchainConnectorConfig

Configuration of the HoprBlockchainConnector.

ChainKeypair

Represents a keypair consisting of a secp256k1 private and public key

ContractAddresses

Holds addresses of all smart contracts.

HoprBlockchainConnector

A connector acting as middleware between the HOPR APIs (see the hopr_api crate) and the Blokli Client API (see the [blokli_client] crate).

SafePayloadGenerator

Payload generator that generates Safe-compliant ABI

TempDbBackend

A backend that is implemented via redb database stored in a temporary file.

Constants

DEFAULT_BLOKLI_URL

Traits

Backend

Represents a cache backend for the connector.

PayloadGenerator

Trait for various implementations of common on-chain transaction payloads generators.

Functions

create_trustful_hopr_blokli_connector

Convenience function to create [HoprBlokliConnector] with contract addresses retrieved from the given client.

create_trustless_hopr_blokli_connector

Convenience function to create [HoprBlokliConnector] with own contract addresses.

init_blokli_connector

Type Aliases

HoprBlockchainSafeConnector

Connector to HOPR on-chain contracts that uses multisig Safe as a signer and TempDbBackend.