hopr_chain_connector

Module api

On-chain operations-related API traits.

Structs

AccountEntry

Represents a node announcement entry on the blockchain. This contains the node’s public key and optional announcement information (multiaddress).

AccountSelector

Selector for on-chain node accounts.

Balance

Represents a non-negative balance of some Currency.

BatchRedemptionResult

Result of [redeem_tickets_via_selector].

ChainInfo

Contains information about the HOPR on-chain network deployment.

ChainPathResolver

[PathAddressResolver] which uses the HOPR chain API to resolve addresses and channels.

ChannelEntry

Overall description of a channel

ChannelSelector

Selector for channels.

ContractAddresses

Holds addresses of all smart contracts.

DeployedSafe

Information about a deployed Safe.

DomainSeparators

Contains domain separator information.

HoprSphinxHeaderSpec

Current Sphinx header specification for the HOPR protocol.

Multiaddr

Representation of a Multiaddr.

RedeemableTicket

Represents a winning ticket that can be successfully redeemed on-chain.

Utc

The UTC time zone. This is the most efficient time zone when you don’t need the local time. It is also used as an offset (which is also a dummy type).

VerifiedTicket

Holds a ticket that has been already verified. This structure guarantees that [Ticket::get_hash()] of VerifiedTicket::verified_ticket() is always equal to VerifiedTicket::verified_hash

WinningProbability

Represents a ticket-winning probability.

Enums

AnnouncementError

Error that can occur when making a node announcement.

ChainEvent

Enumeration of HOPR chain events.

ChannelDirection

Describes a direction of node’s own channel. The direction of a channel that is not own is undefined.

ChannelStatusDiscriminants

SafeRegistrationError

Error that can occur when registering a node with a Safe.

SafeSelector

Selector for deployed Safes.

StateSyncOptions

Indicates if the current state should be emitted in the form of events in the subscription stream.

TicketRedeemError

Errors that can occur during ticket redemption.

Traits

ChainEvents

Allows subscribing to on-chain events.

ChainKeyOperations

Operations for offchain keys.

ChainReadAccountOperations

Chain operations that read on-chain node accounts.

ChainReadChannelOperations

On-chain read operations regarding channels.

ChainReadSafeOperations

Operations for reading Safe information.

ChainValues

Retrieves various on-chain information.

ChainWriteAccountOperations

On-chain write operations regarding on-chain node accounts.

ChainWriteChannelOperations

On-chain write operations regarding channels.

ChainWriteSafeOperations

Operations for writing Safe information.

ChainWriteTicketOperations

On-chain write operations with tickets.

Currency

Represents a general currency - like a token or a coin.

HoprChainApi

Complete set of HOPR on-chain operation APIs.

HoprStaticChainApi

Reduced set of read-only HOPR on-chain operation APIs.

KeyIdMapper

Trait that defines 1:1 mapper between key identifiers and the actual public keys.

Type Aliases

ChainReceipt

Receipt of an on-chain operation.

ChannelId

Alias for the Hash representing a channel ID.

DateTime

HoprBalance

HoprKeyIdent

Type alias for 32-bit HOPR Offchain Public Key Identifier.

HoprSphinxSuite

Currently used public key cipher suite for Sphinx.