hopr_chain_connector

Struct HoprBlockchainConnector

pub struct HoprBlockchainConnector<C, B, P, R> { /* private fields */ }

A connector acting as a middleware between the HOPR APIs (see the hopr_api crate) and the Blokli Client API (see the [blokli_client] crate).

The connector object cannot be cloned, and shall be used inside an Arc if cloning is needed.

Implementations

impl<B, C, P> HoprBlockchainConnector<C, B, P, P::TxRequest>
where B: Backend + Send + Sync + 'static, C: BlokliSubscriptionClient + BlokliQueryClient + BlokliTransactionClient + Send + Sync + 'static, P: PayloadGenerator + Send + Sync + 'static, P::TxRequest: Send + Sync + 'static,

pub fn new( chain_key: ChainKeypair, cfg: BlockchainConnectorConfig, client: C, backend: B, payload_generator: P, ) -> Self

Creates a new instance.

pub async fn connect(&mut self, timeout: Duration) -> Result<(), ConnectorError>

Connects to the chain using the underlying client, syncs all on-chain data and subscribes for all future updates.

If the sync of the current state does not happen within timeout, a ConnectorError::ConnectionTimeout error is returned.

pub fn client(&self) -> &C

Returns the reference to the underlying client.

impl<B, C, P, R> HoprBlockchainConnector<C, R, B, P>

pub fn invalidate_caches(&self)

Invalidates all cached on-chain data.

impl<B, C, P, R> HoprBlockchainConnector<C, B, P, R>
where B: Backend + Send + Sync + 'static, C: Send + Sync, P: Send + Sync, R: Send + Sync,

pub fn as_path_resolver(&self) -> ChainPathResolver<'_, Self>

Returns a PathAddressResolver using this connector.