hopr_chain_connector

Struct HoprBlockchainReader

pub struct HoprBlockchainReader<C>(/* private fields */);

A simplified version of HoprBlockchainConnector which only implements HOPR Chain API partially, allowing for read-only operations.

This object specifically implements only the following traits:

ChainValues

ChainReadSafeOperations

The implementation is currently realized using the Blokli client and acts as a partial HOPR Chain API compatible wrapper for blokli_client::BlokliClient.

This object is useful for bootstrapping purposes that usually precede construction of the full connector.

Implementations

impl<C> HoprBlockchainReader<C>

pub fn new(client: C) -> Self

Creates new instance given the client.