Crate hopr_chain_connector

Modules

api

On-chain operations-related API traits.

blokli_client

Re-exports of the blokli_client crate.

errors

Structs

Address

Represents an Ethereum address

BasicPayloadGenerator

Generates transaction payloads that do not use Safe-compliant ABI

BlockchainConnectorConfig

Configuration of the HoprBlockchainConnector.

ChainKeypair

Represents a keypair consisting of a secp256k1 private and public key

ContractAddresses

Holds addresses of all smart contracts.

HoprBlockchainConnector

A connector acting as middleware between the HOPR APIs (see the [hopr_api] crate) and the Blokli Client API (see the [blokli_client] crate).

HoprBlockchainReader

A simplified version of HoprBlockchainConnector which only implements HOPR Chain API partially, allowing for read-only operations.

SafePayloadGenerator

Payload generator that generates Safe-compliant ABI

TempDbBackend

A backend that is implemented via redb database stored in a temporary file.

Enums

TempDbError

Errors from the temporary database backend.

Traits

Backend

Represents a cache backend for the connector.

PayloadGenerator

Trait for various implementations of common on-chain transaction payloads generators.

Functions

create_trustful_hopr_blokli_connector

Convenience function to create HoprBlockchainConnector with contract addresses retrieved from the given client.

create_trustful_safeless_hopr_blokli_connector

Convenience function to create HoprBlockchainConnector with contract addresses retrieved from the given client.

create_trustless_hopr_blokli_connector

Convenience function to create HoprBlockchainConnector with own contract addresses.

create_trustless_safeless_hopr_blokli_connector

Convenience function to create HoprBlockchainConnector with own contract addresses.

Type Aliases

HoprBlockchainBasicConnector

Connector to HOPR on-chain contracts that uses standard EOA as a signer and TempDbBackend.

HoprBlockchainSafeConnector

Connector to HOPR on-chain contracts that uses multisig Safe as a signer and TempDbBackend.